third_party/blink/web_tests/external/wpt/editing/other/insertparagraph-in-child-of-head.tentative_designMode=off_white-space=pre-expected.txt

This is a testharness.js-based test.
FAIL insertParagraph in empty <title style="display:block;white-space:pre"> should not split it promise_test: Unhandled rejection with value: object "Error: element send_keys intercepted error"
FAIL insertParagraph in <title style="display:block;white-space:pre"> containing text should not split it assert_equals: 1 preformatted linefeed should be inserted expected "<head><title>ab\ncd</title></head><body><br></body>" but got "<head><title>ab</title><title style=\"display:block;white-space:pre\">cd</title></head><body><br></body>"
FAIL insertParagraph in empty <style style="display:block;white-space:pre"> should not split it promise_test: Unhandled rejection with value: object "Error: element send_keys intercepted error"
FAIL insertParagraph in <style style="display:block;white-space:pre"> containing text should not split it assert_equals: 1 preformatted linefeed should be inserted expected "<head><title>iframe</title><style>ab\ncd</style></head><body><br></body>" but got "<head><title>iframe</title><style>ab</style><style style=\"display:block;white-space:pre\">cd</style></head><body><br></body>"
FAIL insertParagraph in empty <script style="display:block;white-space:pre"> should not split it assert_equals: 2 preformatted linefeeds should be inserted expected "<head><title>iframe</title><script>\n\n</script></head><body><br></body>" but got "<head><title>iframe</title><script></script></head><body><br><div><br></div></body>"
FAIL insertParagraph in <script style="display:block;white-space:pre"> containing text should not split it assert_equals: 1 preformatted linefeed should be inserted expected "<head><title>iframe</title><script>// ab\n// cd</script></head><body><br></body>" but got "<head><title>iframe</title><script>// ab</script><script style=\"display:block;white-space:pre\">// cd</script></head><body><br></body>"
FAIL insertParagraph in empty <div style="white-space:pre"> in the <head> should split the <div> promise_test: Unhandled rejection with value: object "Error: element send_keys intercepted error"
PASS insertParagraph in <div style="white-space:pre"> (containing only a <br>) in the <head> should split the <div> element
PASS insertParagraph in <div style="white-space:pre"> (containing text) in the <head> should split the <div> element
Harness: the test ran to completion.
